feat: integrate react toastify
