+++ frontend/src/components/ui/Layout/index.tsx
@@ -1,9 +1,11 @@
 import { PropsWithChildren } from 'react';
 import { Outlet } from 'react-router-dom';
+import { ToastContainer } from 'react-toastify';
 import Header from '../Header';
 import Footer from '../Footer';
-import styles from './Layout.module.scss';
 import Wrapper from '../Wrapper';
+import styles from './Layout.module.scss';
+import 'react-toastify/dist/ReactToastify.css';
 
 const Layout = ({ children }: PropsWithChildren) => (
   <div className={styles.container}>
@@ -15,6 +17,7 @@ const Layout = ({ children }: PropsWithChildren) => (
       </Wrapper>
     </main>
     <Footer />
+    <ToastContainer />
   </div>
 );
