+++ frontend/src/utils/showToastMessage.ts
@@ -1,0 +1,5 @@
+import { ToastOptions, toast } from 'react-toastify';
+
+const showToastMessage = (message: string, options?: ToastOptions) => toast(message, options);
+
+export { showToastMessage };
